% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{abbrv}
\alias{abbrv}
\title{Create Abbreviations}
\usage{
abbrv(x, digits = TRUE)
}
\arguments{
\item{x}{A character vector}

\item{digits}{Include digits in the abbreviation}
}
\value{
Abbreviated versions of \code{x}
}
\description{
Create Abbreviations
}
\examples{
abbrv(c('HelloWorld', 'LetsGo3', 'tomato'))
abbrv(c('HelloWorld', 'LetsGo3', 'tomato'), digits = FALSE)
abbrv('Wow3', digits = FALSE)

}
