% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_theory.R
\name{blend_colors}
\alias{blend_colors}
\title{Blend two or more colors together}
\usage{
blend_colors(..., as.rgb = FALSE)
}
\arguments{
\item{...}{Two or more colors to blend together; these can be in a vector
or standalone}

\item{as.rgb}{Return in RGB form, otherwise return in hexadecimal form}
}
\value{
The blended color in RGB form (1 x 3 matrix) or hexadecimal form
}
\description{
Blend two or more colors together
}
\examples{
blend_colors('red', 'green')
if (requireNamespace('scales', quietly = TRUE)) {
  scales::show_col(c('red', blend_colors('red', 'green'), 'green'), ncol = 3)
}

}
\seealso{
Other color_theory: 
\code{\link{col2hex}()},
\code{\link{contrast-theory}},
\code{\link{text_color}()}
}
\concept{color_theory}
