% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_theory.R
\name{col2hex}
\alias{col2hex}
\title{Convert R colors to hexadecimal}
\usage{
col2hex(...)
}
\arguments{
\item{...}{R colors}
}
\value{
The hexadecimal representations of input colors
}
\description{
Convert R colors to hexadecimal
}
\examples{
col2hex('black', 'red', 'grey')

}
\seealso{
Other color_theory: 
\code{\link{blend_colors}()},
\code{\link{contrast-theory}},
\code{\link{text_color}()}
}
\concept{color_theory}
