% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_theory.R
\name{contrast-theory}
\alias{contrast-theory}
\alias{intensity}
\alias{luminance}
\title{Color Intensity}
\source{
\url{https://stackoverflow.com/questions/3942878/how-to-decide-font-color-in-white-or-black-depending-on-background-color}
}
\usage{
intensity(color)

luminance(color)
}
\arguments{
\item{color}{A vector of colors}
}
\value{
A vector of intensities/luminances for each color
}
\description{
Get the intensity and/or luminance of a color
}
\examples{
intensity(color = c('black', 'white', '#E76BF3'))

luminance(color = c('black', 'white', '#E76BF3'))

}
\seealso{
Other color_theory: 
\code{\link{blend_colors}()},
\code{\link{col2hex}()},
\code{\link{text_color}()}
}
\concept{color_theory}
