% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{interleave}
\alias{interleave}
\title{Interleave vectors together}
\usage{
interleave(...)
}
\arguments{
\item{...}{Vectors to be interleaved}
}
\value{
A vector with the values from each vector in ... interleaved
}
\description{
Interleave vectors together
}
