% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_data_frame}
\alias{is_data_frame}
\alias{is_bare_data_frame}
\title{Is an Object a Data Frame}
\usage{
is_data_frame(x)

is_bare_data_frame(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if \code{x} is a data frame, otherwise \code{FALSE}
}
\description{
Is an Object a Data Frame
}
