% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_named_list}
\alias{is_named_list}
\title{Is a List Named}
\usage{
is_named_list(x, pass.zero = FALSE)
}
\arguments{
\item{x}{A list}

\item{pass.zero}{Return \code{TRUE} for zero-length lists}
}
\value{
...
}
\description{
Is a List Named
}
