% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyjs.R
\name{js_tab_hide}
\alias{js_tab_hide}
\title{Hide Shiny Tabs with \pkg{ShinyJS}}
\usage{
js_tab_hide(id, values, fxn = "hide")
}
\arguments{
\item{id}{ID of a \code{\link[shiny]{tabsetPanel}}}

\item{values}{One or more values of a \code{\link[shiny]{tabPanel}}
(see the \code{value} parameter)}

\item{fxn}{Name of JavaScript call function}
}
\value{
A string with a JavaScript function to hide a set of tabs
}
\description{
Hide Shiny Tabs with \pkg{ShinyJS}
}
\note{
This function is designed to run custom JavaScript code using
\code{\link[shinyjs:extendShinyJS]{shinyjs::extendShinyJS}()}; use of
custom JavaScript code requires the \pkg{V8} package.
\pkg{V8} requires a local install of either the
\href{https://chromium.googlesource.com/v8/v8}{V8} JavaScript Engine or
\href{https://nodejs.org/en/}{Node.js}
}
\seealso{
\code{\link[shinyjs:extendShinyJS]{shinyjs::extendShinyJS}()}
\code{\link[shiny:tabPanel]{shiny::tabPanel}()}

Other shiny: 
\code{\link{js_tab_key}()}

Other shinyjs: 
\code{\link{js_tab_key}()}
}
\concept{shiny}
\concept{shinyjs}
