% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyjs.R
\name{js_tab_key}
\alias{js_tab_key}
\title{Get JavaScript IDs for Shiny Tabs}
\usage{
js_tab_key(id, values)
}
\arguments{
\item{id}{ID of a \code{\link[shiny]{tabsetPanel}}}

\item{values}{One or more values of a \code{\link[shiny]{tabPanel}}
(see the \code{value} parameter)}
}
\value{
A string with the JavaScript ID for a given set of tabs
}
\description{
Get JavaScript IDs for Shiny Tabs
}
\seealso{
\code{\link[shiny:tabsetPanel]{shiny::tabsetPanel}()}
\code{\link[shiny:tabPanel]{shiny::tabPanel}()}

Other shiny: 
\code{\link{js_tab_hide}()}

Other shinyjs: 
\code{\link{js_tab_hide}()}
}
\concept{shiny}
\concept{shinyjs}
