% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mamisc-package}
\alias{mamisc}
\alias{mamisc-package}
\title{mamisc: Miscellaneous Functions by Paul Hoffman}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mojaveazure/mamisc}
  \item Report bugs at \url{https://github.com/mojaveazure/mamisc/issues}
}

}
\author{
\strong{Maintainer}: Paul Hoffman \email{hoff0792@umn.edu} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID})

}
