% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{ns_pkg_imports}
\alias{ns_pkg_imports}
\title{Find Imported Packages}
\usage{
ns_pkg_imports(path = ".")
}
\arguments{
\item{path}{\verb{[character(1)]}\cr
The start directory.}
}
\value{
A character vector with all imported packages
}
\description{
Find Imported Packages
}
\note{
\Sexpr[stage=build]{mamisc::rd_required_pkgs('rprojroot')}
}
\keyword{internal}
