% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/www.R
\name{online}
\alias{online}
\title{Is a Resource Online and Available}
\usage{
online(uri, strict = FALSE, timeout = getOption(x = "timeout"))
}
\arguments{
\item{uri}{The URI of a given resource}

\item{strict}{Ensures the HTTP status code is \code{200}}

\item{timeout}{Timeout in seconds; defaults to the option
\dQuote{\code{timeout}} (currently set to
\Sexpr[stage=render]{getOption("timeout")} seconds)}
}
\value{
\code{TRUE} if \code{uri} is accessible, otherwise \code{FALSE}
}
\description{
Check to see if a given resource is online and accessible over the internet
}
\note{
\Sexpr[stage=build]{mamisc::rd_required_pkgs('httr')}
}
