% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{powerset}
\alias{powerset}
\title{Generate a powerset}
\usage{
powerset(x)
}
\arguments{
\item{x}{A list or vector of values to generate all possible combinations of}
}
\value{
A list with all possible combinations of x
}
\description{
Generate a powerset
}
\examples{
powerset(x = 1:3)
powerset(x = letters[1:3])

}
\keyword{utilities}
