% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{rd_required_pkgs}
\alias{rd_required_pkgs}
\title{Rdocumentation Note of Required Packages}
\usage{
rd_required_pkgs(x, multiple = FALSE)
}
\arguments{
\item{x}{A vector of package names}

\item{multiple}{...}
}
\value{
...
}
\description{
Rdocumentation Note of Required Packages
}
\examples{
#' \Sexpr[results=rd]{odds::rd_required_pkgs()}
}
\seealso{
Other roxygen: 
\code{\link{rd_ilist}()}
}
\concept{roxygen}
