% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{is_bare_list}
\alias{is_na}
\alias{is_named}
\alias{is_named2}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rlang}{\code{\link[rlang:bare-type-predicates]{is_bare_list}}, \code{\link[rlang:are_na]{is_na}}, \code{\link[rlang]{is_named}}, \code{\link[rlang:is_named]{is_named2}}}
}}

